Extract getResultPath into its own module
- Move getResultPath and ADMIN_SECTIONS out of SearchResultsPanel.tsx
  into src/components/search/getResultPath.ts so the component file only
  exports components, clearing the react-refresh/only-export-components
  oxlint warning.
- Keep ADMIN_SECTIONS module-private (no external consumers) to satisfy
  the fallow unused-exports gate.
- Point the test import at the new module and add a Release-without-
  project_id null case mirroring the existing Document case.
