+++ src/components/search/SearchResultsPanel.tsx
@@ -9,6 +9,7 @@ import {
   getConfidenceLabel,
   type SearchResult,
 } from '@/api/search'
+import { getResultPath } from '@/components/search/getResultPath'
 import { Collapsible, CollapsibleContent } from '@/components/ui/collapsible'
 import { Label } from '@/components/ui/label'
 import { Slider } from '@/components/ui/slider'
@@ -397,34 +398,6 @@ export function SearchResultsPanel({
       </div>
     </div>
   )
-}
-
-// Maps a node label to its /admin section. These node types are edited on
-// the admin pages, which route by slug (see Admin.tsx).
-const ADMIN_SECTIONS: Record<string, string> = {
-  Blueprint: 'blueprints',
-  DocumentTemplate: 'document-templates',
-  Environment: 'environments',
-  LinkDefinition: 'link-definitions',
-  Organization: 'organizations',
-  ProjectType: 'project-types',
-  Role: 'roles',
-  Team: 'teams',
-  ThirdPartyService: 'third-party-services',
-}
-
-export function getResultPath(result: SearchResult): null | string {
-  const { node_id, node_label, project_id, slug } = result
-  if (node_label === 'Project') return `/projects/${node_id}`
-  // Documents and Releases live under their parent project; project_id is
-  // supplied by the search API's result enrichment.
-  if (node_label === 'Document')
-    return project_id ? `/projects/${project_id}/documents/${node_id}` : null
-  // ReleasesTab does not yet consume the subId, so this lands on the tab.
-  if (node_label === 'Release')
-    return project_id ? `/projects/${project_id}/releases/${node_id}` : null
-  const section = ADMIN_SECTIONS[node_label]
-  return section && slug ? `/admin/${section}/${slug}` : null
 }
 
 function highlightKeywords(text: string, query: string): React.ReactNode {
